Actually, rename {Vdom ⇒ React}Fragment
