+++ core/src/main/scala/japgolly/scalajs/react/vdom/VdomNode.scala
@@ -1,6 +1,6 @@
 package japgolly.scalajs.react.vdom
 
-import japgolly.scalajs.react.{Callback, Key, raw => Raw}
+import japgolly.scalajs.react.{Callback, raw => Raw}
 import scala.scalajs.js
 
 sealed class VdomNode(val rawNode: Raw.React.Node) extends TagMod {
@@ -56,23 +56,4 @@ object VdomArray {
   /** Elements must have keys. Array itself cannot. */
   def apply(ns: VdomNode*): VdomArray =
     empty() ++= ns
-}
-
-// =====================================================================================================================
-
-object VdomFragment {
-
-  /** Elements keys are optional. */
-  def apply(ns: VdomNode*): VdomElement =
-    create(null, ns: _*)
-
-  /** Elements keys are optional. */
-  def withKey(key: Key)(ns: VdomNode*): VdomElement = {
-    val jsKey: Raw.React.Key = key
-    val props = js.Dynamic.literal("key" -> jsKey.asInstanceOf[js.Any])
-    create(props, ns: _*)
-  }
-
-  private def create(props: js.Object, ns: VdomNode*): VdomElement =
-    VdomElement(Raw.React.createElement(Raw.React.Fragment, props, ns.map(_.rawNode): _*))
 }
