+++ doc/changelog/1.2.0.md
@@ -8,7 +8,7 @@ The largest new feature in this release is React 16 support (and adherence).
   * `EmptyVdom`
   * `Option[_]`
   * `VdomArray`
-  * `VdomFragment`
+  * `ReactFragment`
 
 * [React 16] The `.raw` packages have been migrated to the new types of React 16 and reorganised to match their JS counterparts.
   Examples:
