+++ test/src/test/scala/japgolly/scalajs/react/core/vdom/VdomTest.scala
@@ -18,20 +18,20 @@ object VdomTest extends TestSuite {
         assertRender(comp(), exp)
       }
 
-      'byte      - test(50: Byte,                                      "50")
-      'short     - test(45: Short,                                     "45")
-      'int       - test(666,                                           "666")
-      'long      - test(123L,                                          "123")
-      'double    - test(12.3,                                          "12.3")
-      'string    - test("yo",                                          "yo")
-      'empty     - test(EmptyVdom,                                     "")
-      'optionN   - test(Option.empty[Int],                             "")
-      'optionS   - test(Option(13),                                    "13")
-      'optionSN  - test(Option(Option.empty[Int]),                     "")
-      'optionSS  - test(Option(Option(13)),                            "13")
-      'vdomArray - test(VdomArray("hehe", <.div(^.key := 1, "one")),   "hehe<div>one</div>") // needs keys apparently
-      'fragment  - test(VdomFragment("hehe", <.div("one")),            "hehe<div>one</div>") // keys not required
-      'fragmentK - test(VdomFragment.withKey(1)("hehe", <.div("one")), "hehe<div>one</div>") // keyABLE
+      'byte      - test(50: Byte,                                       "50")
+      'short     - test(45: Short,                                      "45")
+      'int       - test(666,                                            "666")
+      'long      - test(123L,                                           "123")
+      'double    - test(12.3,                                           "12.3")
+      'string    - test("yo",                                           "yo")
+      'empty     - test(EmptyVdom,                                      "")
+      'optionN   - test(Option.empty[Int],                              "")
+      'optionS   - test(Option(13),                                     "13")
+      'optionSN  - test(Option(Option.empty[Int]),                      "")
+      'optionSS  - test(Option(Option(13)),                             "13")
+      'vdomArray - test(VdomArray("hehe", <.div(^.key := 1, "one")),    "hehe<div>one</div>") // needs keys apparently
+      'fragment  - test(ReactFragment("hehe", <.div("one")),            "hehe<div>one</div>") // keys not required
+      'fragmentK - test(ReactFragment.withKey(1)("hehe", <.div("one")), "hehe<div>one</div>") // keyABLE
 
       'booleanF  - compileError("""test(false, "")""")
       'booleanT  - compileError("""test(true, "")""")
